move config/fileutils to new package
